[SimpleSanction] Use Modals to request the sanction duration and the reason (only if buttons are used).

+++ simplesanction/types.py
@@ -37,50 +37,75 @@ class Action:
             if not v.startswith("_") and v not in ["to_json", "process"]
         }
 
-    async def process(self, ctx: commands.Context, member: discord.Member, duration: typing.Optional[str] = None, reason: typing.Optional[str] = "The reason was not given.", finish_message_enabled: typing.Optional[bool] = True, reason_required: typing.Optional[bool] = True, confirmation: typing.Optional[bool] = False, show_author: typing.Optional[bool] = True, fake_action: typing.Optional[bool] = False) -> bool:
+    async def process(self, ctx: commands.Context, interaction: typing.Optional[discord.Interaction], member: discord.Member, duration: typing.Optional[str] = None, reason: typing.Optional[str] = "The reason was not given.", finish_message_enabled: typing.Optional[bool] = True, reason_required: typing.Optional[bool] = True, confirmation: typing.Optional[bool] = False, show_author: typing.Optional[bool] = True, fake_action: typing.Optional[bool] = False) -> bool:
         if (await self.cog.config.guild(ctx.guild).use_warn_system()) and self.warn_system_command is not None and ctx.bot.get_cog("WarnSystem") is not None:
             use_warn_system = True
         else:
             use_warn_system = False
             if self.cog_required is not None and not ctx.bot.get_cog(self.cog_required):
+                if interaction is not None:
+                    await interaction.response.defer()
                 await ctx.send(
                     _(
                         "The cog `{cog_required}` is not loaded. Please load it, with the command `{prefix}load {cog_required_lowered}`."
                     ).format(prefix=ctx.prefix, cog_required=self.cog_required, cog_required_lowered=self.cog_required.lower())
                 )
                 return False
-        if duration is None and self.duration_ask_message is not None:
-            duration_message = await ctx.send(_(self.duration_ask_message).format(member=member, duration=str(parse_timedelta(duration)) if duration is not None else None, reason=reason, channel=ctx.channel))
-            try:
-                pred = MessagePredicate.same_context(ctx)
-                msg = await ctx.bot.wait_for("message", timeout=60, check=pred)
-                await self.cog.cogsutils.delete_message(duration_message)
-                await self.cog.cogsutils.delete_message(msg)
-                if msg.content.lower() == "cancel":
-                    return
-                duration = msg.content
-            except asyncio.TimeoutError:
-                await ctx.send(_("Timed out, please try again."))
-                return
-        if reason is None:
-            if self.reason_ask_message is not None and reason_required:
-                reason_message = await ctx.send(_(self.reason_ask_message).format(member=member, duration=str(parse_timedelta(duration)) if duration is not None else None, reason=reason, channel=ctx.channel))
+
+        if interaction is not None:
+            extra_params_inputs = {}
+            if duration is None and self.duration_ask_message is not None:
+                extra_params_inputs["duration"] = discord.ui.TextInput(label="Duration", style=discord.TextStyle.short, required=True, placeholder=_("The sanction duration."))
+            if reason is None and self.reason_ask_message is not None:
+                extra_params_inputs["reason"] = discord.ui.TextInput(label="Reason", style=discord.TextStyle.paragraph, required=False, placeholder=_("The reason was not given."))
+            if extra_params_inputs:
+                modal = discord.ui.Modal(title=f"{self.emoji} {self.label}", custom_id="SimpleSanction")
+                modal.on_submit = lambda modal_interaction: modal_interaction.response.defer()
+                [modal.add_item(text_input) for text_input in extra_params_inputs.values()]
+                await interaction.response.send_modal(modal)
+                if await modal.wait():
+                    return  # timeout
+                if duration is None and self.duration_ask_message is not None:
+                    duration = extra_params_inputs["duration"].value
+                if reason is None and self.reason_ask_message is not None:
+                    reason = extra_params_inputs["reason"].value or _("The reason was not given.")
+            else:
+                await interaction.response.defer()
+        else:
+            if duration is None and self.duration_ask_message is not None:
+                duration_message = await ctx.send(_(self.duration_ask_message).format(member=member, duration=str(parse_timedelta(duration)) if duration is not None else None, reason=reason, channel=ctx.channel))
                 try:
                     pred = MessagePredicate.same_context(ctx)
                     msg = await ctx.bot.wait_for("message", timeout=60, check=pred)
-                    await self.cog.cogsutils.delete_message(reason_message)
+                    await self.cog.cogsutils.delete_message(duration_message)
                     await self.cog.cogsutils.delete_message(msg)
                     if msg.content.lower() == "cancel":
                         return
-                    reason = _("The reason was not given.") if reason == "not" else msg.content
+                    duration = msg.content
                 except asyncio.TimeoutError:
                     await ctx.send(_("Timed out, please try again."))
                     return
-            else:
-                reason = _("The reason was not given.")
+            if reason is None:
+                if self.reason_ask_message is not None and reason_required:
+                    reason_message = await ctx.send(_(self.reason_ask_message).format(member=member, duration=str(parse_timedelta(duration)) if duration is not None else None, reason=reason, channel=ctx.channel))
+                    try:
+                        pred = MessagePredicate.same_context(ctx)
+                        msg = await ctx.bot.wait_for("message", timeout=60, check=pred)
+                        await self.cog.cogsutils.delete_message(reason_message)
+                        await self.cog.cogsutils.delete_message(msg)
+                        if msg.content.lower() == "cancel":
+                            return
+                        reason = _("The reason was not given.") if reason == "not" else msg.content
+                    except asyncio.TimeoutError:
+                        await ctx.send(_("Timed out, please try again."))
+                        return
+                else:
+                    reason = _("The reason was not given.")
+
         if not confirmation and self.confirmation_ask_message is not None and not await self.cog.cogsutils.ConfirmationAsk(ctx, content=_(self.confirmation_ask_message).format(member=member, duration=str(parse_timedelta(duration)) if duration is not None else None, reason=reason, channel=ctx.channel)) and not ctx.assume_yes:
             await self.cog.cogsutils.delete_message(ctx.message)
             return
+
         if finish_message_enabled and self.finish_message is not None:
             embed: discord.Embed = discord.Embed()
             embed.title = f"Sanction Member - {self.emoji} {self.label}"
@@ -124,6 +149,7 @@ class Action:
             finish_message = await ctx.send(embed=embed)
         else:
             finish_message = None
+
         if not fake_action:
             command = self.warn_system_command if use_warn_system else self.command
             command = command.format(member=member, duration=str(parse_timedelta(duration)) if duration is not None else None, reason=reason, channel=ctx.channel)
@@ -140,6 +166,7 @@ class Action:
                 raise commands.UserFeedbackCheckFailure(
                     _("You can't execute this command, in this context.")
                 )
+
         if finish_message is not None:
             try:
                 await finish_message.add_reaction("✅")
